moving to sensor based approach
