+++ src/view/drag-handle/is-sloppy-click-threshold-exceeded.js
@@ -1,0 +1,9 @@
+// @flow
+import type { Position } from '../../types';
+// The amount of pixels that need to move before we consider the movement
+// a drag rather than a click.
+export const sloppyClickThreshold: number = 5;
+
+export default (original: Position, current: Position): boolean =>
+  Math.abs(current.x - original.x) >= sloppyClickThreshold ||
+  Math.abs(current.y - original.y) >= sloppyClickThreshold;
