+++ src/view/drag-handle/stop-event.js
@@ -1,0 +1,5 @@
+// @flow
+export default (event: Event) => {
+  event.preventDefault();
+  event.stopPropagation();
+};
